% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_var.R
\name{create_var}
\alias{create_var}
\title{Create a multivariate time series from a VAR}
\usage{
create_var(coefficient_matrices, Sigma_a, n, seed = 56143868)
}
\arguments{
\item{x}{Some sort of input}
}
\value{
A character string
}
\description{
Create a multivariate time series from a VAR
}
\examples{
create_var("Hello")
}
